requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==1.9
gevent==1.3.0
greenlet==0.4.13
grequests==0.3.0
idna==2.6
pkginfo==1.4.2
pycparser==2.18
PyJWT==1.6.1
pypandoc==1.4
requests==2.18.4
requests-toolbelt==0.8.0
six==1.11.0
tqdm==4.23.3
twine==1.11.0
urllib3==1.22
